Exposed audio system to script via proxy class
A new proxy class, AudioServiceScriptProxy, exposes an instance of
AudioService as a set of static methods. This allows easy binding of the
static methods to the script layer without having to bind the
AudioService class directly, and avoids weird construction issues with
Sqrat. If no AudioService is set then the methods do nothing, so the
scripts still work fine.