Made state changes safer by using request/return
The Hero's state machine change mechanism was dangerous if you change
state in the middle of a method but forget to return immediately because
the calling object is destroyed instantly... which is bad. Now to change
states you're supposed to request a state change and then return the
correct "state change action" from your update function. This allows the
"state machine mechanism" to handle the transition.