Fix camera scrolling behavior
Made camera scrolling behavior work like it does in Mega Man 3. The
camera pans left or right when Rock passes through the mid-line of the
screen. You can be offset and it will still follow you, but it won't
auto-center. It now only moves when Rock moves a full pixel so this
should make things look nice and crisp as it moves.